requirements.txt
matplotlib == 3.9.4
pandas == 2.3.1
pip == 25.1.1
PySocks == 1.7.1
python-barcode == 0.15.1
selenium == 4.34.2
setuptools == 58.1.0
webdriver-manager == 4.0.2
xlsxwriter == 3.2.5